refactor(server): relocate the prompt parser and connector filter
Two small server-package helpers move to the packages they belong to:

  - the OIDC prompt parser (Prompt/ParsePrompt) moves to server/oauth2, next to
    the other request-parameter parsing.
  - filterConnectors becomes connectors.Filter, alongside the ConnectorAllowed /
    GrantTypeAllowed policy it complements.
