+++ server/oauth2/prompt.go
@@ -1,4 +1,4 @@
-package server
+package oauth2
 
 import (
 	"fmt"
