+++ server/oauth2/prompt_test.go
@@ -1,4 +1,4 @@
-package server
+package oauth2
 
 import (
 	"testing"
